+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.97.76", date: "2026-03-27" };
+  return { version: "8.97.77", date: "2026-03-27" };
 }